pyproject.toml
[project]
name = "second-streamlit-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-adk>=1.5.0",
    "streamlit>=1.46.1",
]
